feat: day 12
Day 12 is interesting. Part 1 was pretty easy but part 2 required some clever thinking.
I ended up realizing that edges should be calculated from the plots and not from the outer neighboring plots.
To make sure each edge is only counted once, I only count the top-most or left-most plot in an edge,
that concept was the fundamental idea of the pt. 2 solution.